Do not ignore subfolder with name equal to root.
Before this commit, a subfolder with the same name as the root/source
folder would be ignored from the coverage reports (see
https://github.com/mozilla/grcov/issues/483). This was occurring because
the ancestor of the subfolder was being pruned (because it matched the
end of the source folder) in the method used to guess the absolute path.
This resulted in an incorrect guess of the absolute path.

With this commit, we first check whether the join of the subfolder
relative path with the source folder exists, and if so, return it as the
guess of the absolute path. If the join doesn't exist, we use the
previous guess method.